demo_tts_data_factory/requirements.txt
pydub>=0.25,<1
PyYAML>=6.0,<7
requests>=2.32,<3
soundfile>=0.12,<1
typer>=0.12,<1
